requirements.txt
albumentations==1.3.1
click==8.1.7
numpy==1.26.4
Pillow==9.3.0
Pillow==10.2.0
pytorch_lightning==2.0.9.post0
PyYAML==6.0.1
torch==2.1.1+cu118
torchmetrics==1.2.0
torchvision==0.16.1+cu118
